Tailscale up: allow flags with no bool value
This change makes it so flags can be specified (in hiera) with no
specific truth value. This is required for flags such as
--advertise-exit-node or --ssh, which do not allow flag='' as their
specification.

Previously, specifying
  --ssh=true
would cause tailscale to print the help message and exit. This change
makes it possible to specify just the '--ssh' part.

In hiera, this can be accomplished by specifying just the key with no
value, i.e.

tailscale::up_options:
  ssh:

(note no characters on the line after the ':' in 'ssh:')

There's a Boolean type check in this, which was necessary to prevent
puppet from interpreting hiera like `ssh: "false"` as true. (Puppet
documentation indicates that the string "false" is truthy.)